Rename var to improve clarity
A parent of X denotes a the immediate progenitor of X. Using `hasSelectedInnerBlock` with the `deep` argument checks all levels of nesting of InnerBlocks. Therefore the value could be the parent but it could also be a grand-parent (and so on). As a result a more useful term is “ancestor”.